requirements.txt
# Python 3.11 required
#
# Installation tends to be finicky, audiocraft hard-pins torch==2.1.0 which conflicts with
# xformers 0.0.35, so please install it separately with --no-deps:
#
#   Step 1:  pip install -r requirements.txt
#   Step 2:  pip install --no-deps "audiocraft @ git+https://github.com/facebookresearch/audiocraft.git"

--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128

xformers==0.0.35

av>=10.0.0
demucs>=4.0.0
einops>=0.6.1
encodec>=0.1.1
flashy>=0.0.2
hydra-core>=1.3.2
hydra_colorlog>=1.2.0
julius>=0.2.7
num2words>=0.5.12
pesq>=0.0.4
protobuf>=3.20.0
pystoi>=0.3.3
spacy>=3.6.0
torchdiffeq>=0.2.3
torchmetrics>=0.11.4
torchtext>=0.17.0

datasets>=2.14.0
huggingface_hub>=0.17.0

frechet-audio-distance>=0.3.0

transformers>=4.35.2

librosa>=0.10.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
tqdm>=4.65.0
omegaconf>=2.3.0
python-dotenv>=1.0.0
sentencepiece>=0.1.99
